+++ repo/src/main/java/com/parashift/onlyoffice/constants/Type.java
@@ -3,5 +3,6 @@ package com.parashift.onlyoffice.constants;
 public enum Type {
     WORD,
     CELL,
-    SLIDE
+    SLIDE,
+    FORM
 }